Remove wrapper impl
